+++ app/policies/project_policy.rb
@@ -1,0 +1,27 @@
+class ProjectPolicy < ApplicationPolicy
+  def index?
+    true
+  end
+
+  def show?
+    true
+  end
+
+  def create?
+    !beis_user?
+  end
+
+  def update?
+    !beis_user?
+  end
+
+  def destroy?
+    false
+  end
+
+  class Scope < Scope
+    def resolve
+      scope.all
+    end
+  end
+end