+++ CHANGELOG.md
@@ -45,3 +45,5 @@
 - Users can add budgets to project level activities
 - Users can view Budgets in the Activity XML
 - Consolidate all user roles into 'administrator', removing 'fund_manager' and 'delivery_partner'. All users can temporarily do anything
+- Users who belong to delivery partners see a list of programmes on their organisation page
+- Users are authorised based on the organisations they belong to